% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_papers.R
\name{cleanTexts}
\alias{cleanTexts}
\title{clean pdfs for analysis}
\usage{
cleanTexts(paper_texts, remove_post_text = T, remove_stop_words = T)
}
\arguments{
\item{paper_texts}{a list of paper texts, can be imported using \code{\link{importPDFs}}}

\item{remove_post_text}{logical, if true all text after (and including) acknowledgements,
bibliograpy and references will be removed (but only if only said once)}
}
\value{
tibble with one row per paper and columns \code{paper_name} with names from list
\code{paper_texts}; \code{original_paper_text} with the original text from pdf;
\code{cleaned_paper_text} with a cleaned version of the paper text ready for text mining
analysis; and \code{cleaned_paper_text_lower_case} with a lower case version of the cleaned
paper text.
}
\description{
Cleans some common features from pdf texts
}
