% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikiScraper.R
\name{getElementText}
\alias{getElementText}
\title{Find an element and get text}
\usage{
getElementText(remDr, using, path)
}
\value{
a value containing the text from the element
}
\description{
Find an element and get text
}
