% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_papers.R
\name{importPDFs}
\alias{importPDFs}
\title{import PDFs}
\usage{
importPDFs(pdf_location)
}
\arguments{
\item{pdf_location}{path to folder containing pdf files to be imported}
}
\value{
a list with one element per pdf, element names are the original file names,
within each element, there is an object per page
}
\description{
import all pdfs in a folder
}
