% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances.R
\name{instancesNoOverlap}
\alias{instancesNoOverlap}
\title{INSTANCESNOOVERLAP}
\usage{
instancesNoOverlap(texts, search_string, text_ids, search_id)
}
\arguments{
\item{texts}{vector of texts to be searched within}

\item{search_string}{vector of strings to search}

\item{text_ids}{vector of ids corresponding to texts in \code{x}}

\item{search_id}{vector of ids corresponding to search terms in \code{search_string}}
}
\value{
matrix with number of instances of each search term in each text
}
\description{
as instances but removes all cases where one search term is returned inside another search term
e.g. if 'owl' and 'scops owl' are both search terms, any time scops owl is written, it will not count as an instance of owl
it returns a matrix with number of instances of each search term in each requires stringr to work
}
