% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_papers.R
\name{matchPDFsToRefs}
\alias{matchPDFsToRefs}
\title{Match pdfs with their references}
\usage{
matchPDFsToRefs(PDFs, references_file)
}
\arguments{
\item{PDFs}{a list containing the text for one pdf per element, can be created with
\code{\link{importPDFs}}}

\item{references_file}{file path to .RIS file containing references}
}
\value{
tibble with references and title of PDF they've been matched to
}
\description{
Read pdfs as text and associate with references, pdfs must have been saved with
paper title within file title
}
