% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_papers.R
\name{removePostText}
\alias{removePostText}
\title{remove post text}
\usage{
removePostText(paper_text)
}
\arguments{
\item{paper_text}{character vector with one paper text per element}
}
\value{
paper text with any post text removed
}
\description{
remove all text after (and including) acknowledgements, bibliograpy and references (but only if only said once)
}
