% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikiScraper.R
\name{splitWikiScrape}
\alias{splitWikiScrape}
\title{Find an element and get text}
\usage{
splitWikiScrape(text, contents)
}
\value{
a value containing the text from the element
}
\description{
Find an element and get text
}
