% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikiScraper.R
\name{textListToTibble}
\alias{textListToTibble}
\title{Find an element and get text}
\usage{
textListToTibble(list, directory, name, internalTaxonId, scientificName)
}
\value{
a value containing the text from the element
}
\description{
Find an element and get text
}
