% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikiScraper.R
\name{wikiScrape}
\alias{wikiScrape}
\title{scrape all text for a vector of species from wikipedia}
\usage{
wikiScrape(species)
}
\arguments{
\item{species}{a list of species names to be searched in wikipedia}
}
\value{
a list containing the text from each section of the species' wikipedia
page
}
\description{
scrape all text for a vector of species from wikipedia
}
